scrapyd/requirements.txt
scrapy==2.6.2
scrapyd==1.3.0
psycopg2-binary
furl
python-dotenv==0.19.0
python-decouple==3.5
logparser
setuptools
twisted==21.7.0
pyopenssl==22.0.0
cryptography==38.0.4